pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arxiv-pipeline"
version = "0.1.0"
description = "ArXiv Prefect Pipeline - Backend processing pipeline for arXiv papers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.1.0",
    "minio>=7.2.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "prefect>=2.0",
    "prefect-aws>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "setfit>=1.0.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
]

[tool.setuptools.packages.find]
where = ["src"]
